Allow separating index storage from graph storage
This lets us keep the indexes in a different directory than the graphs,
so I can test code that changes indexing without replacing the existing
indexes.

This commit was produced by vigorous badgering of Anthropic Claude,
which remembered that the benchmarks also needed to be updated, but had
to be remineded about not leaving undefined Snakemake input wildcards
and about not just duplicating your whole function to change one
constant. It took about 3 times as long to do the implementation as I
guessed I would have taken, but the amount of time I then spend fixing
new bugs has yet to be measured.